requirements.txt
# ML Scheduler Project Dependencies

# Core ML libraries
scikit-learn>=1.5.0
xgboost>=2.1.0
torch>=2.4.0
numpy>=1.26.0
pandas>=2.2.0

# MLOps and experiment tracking
mlflow>=2.14.0
feast>=0.40.0
kubeflow-pipelines>=2.2.0

# Kubernetes integration
kubernetes>=30.1.0
prometheus-client>=0.20.0

# Data processing and analysis
scipy>=1.13.0
matplotlib>=3.9.0
seaborn>=0.13.0
plotly>=5.22.0

# Feature engineering
category-encoders>=2.6.0
feature-engine>=1.8.0

# Reinforcement learning
gymnasium>=0.29.0
stable-baselines3>=2.3.0

# Anomaly detection
pyod>=2.0.0

# Data validation and monitoring
great-expectations>=0.18.0
evidently>=0.4.0

# Utilities
pyyaml>=6.0.0
requests>=2.32.0
redis>=5.0.0
python-dateutil>=2.9.0

# Development and testing
pytest>=8.2.0
pytest-cov>=5.0.0
black>=24.4.0
flake8>=7.1.0
mypy>=1.10.0

# Jupyter notebook environment
jupyterlab>=4.2.0
ipywidgets>=8.1.0
notebook>=7.2.0